organize useQuery calls
